+++ R/radae.R
@@ -65,18 +65,13 @@ radae <- function(ADSL, # nolint
                     ),
                   cached = FALSE) {
 
-  stopifnot(is_logical_single(cached))
+  checkmate::assert_flag(cached)
   if (cached) return(get_cached_data("cadae"))
 
   stopifnot(is.data.frame(ADSL))
-  stopifnot(is_integer_single(max_n_aes))
-  stopifnot(is.null(seed) || is_numeric_single(seed))
-  stopifnot(
-    (
-      is_numeric_single(na_percentage) &&
-        na_percentage >= 0 && na_percentage < 1
-      ) || is.na(na_percentage)
-    )
+  checkmate::assert_integer(max_n_aes, len = 1, any.missing = FALSE)
+  checkmate::assert_numeric(seed, null.ok = TRUE, len = 1, any.missing = FALSE)
+  checkmate::assert_numeric(na_percentage, len = 1, any.missing = TRUE, lower = 0, upper = 1)
 
   lookup_ae <- if_null( # nolint
     lookup,