+++ R/raddv.R
@@ -37,14 +37,14 @@ raddv <- function(ADSL, # nolint
                   ),
                   cached = FALSE) {
 
-  stopifnot(is_logical_single(cached))
+  checkmate::assert_flag(cached)
   if (cached) return(get_cached_data("caddv"))
 
   stopifnot(is.data.frame(ADSL))
-  stopifnot(is_integer_single(max_n_dv))
-  stopifnot(is_numeric_single(p_dv) && p_dv > 0 && p_dv <= 1)
-  stopifnot(is.null(seed) || is_numeric_single(seed))
-  stopifnot((is_numeric_single(na_percentage) && na_percentage >= 0 && na_percentage < 1) || is.na(na_percentage))
+  checkmate::assert_integer(max_n_dv, len = 1, any.missing = FALSE)
+  checkmate::assert_numeric(p_dv, len = 1, any.missing = FALSE, lower = 0, upper = 1)
+  checkmate::assert_numeric(seed, null.ok = TRUE, len = 1, any.missing = FALSE)
+  checkmate::assert_numeric(na_percentage, len = 1, any.missing = TRUE, lower = 0, upper = 1)
 
   if (!is.null(seed)) set.seed(seed)