+++ R/utils.R
@@ -4,7 +4,7 @@
 #'
 #' @noRd
 get_cached_data <- function(dataname) {
-  stopifnot(is_character_single(dataname))
+  checkmate::assert_string(dataname)
   if (!("package:random.cdisc.data" %in% search())) {
     stop("cached data can only be loaded if the random.cdisc.data package is attached.",
          "Please run library(random.cdisc.data) before loading cached data.", call. = FALSE)
@@ -29,7 +29,7 @@ get_cached_data <- function(dataname) {
 #' random.cdisc.data:::sample_fct(letters[1:3], 10)
 #' random.cdisc.data:::sample_fct(iris$Species, 10)
 sample_fct <- function(x, N, ...) { # nolint
-  stopifnot(is_numeric_single(N))
+  checkmate::assert_numeric(N, len = 1, any.missing = FALSE)
 
   factor(sample(x, N, replace = TRUE, ...), levels = if (is.factor(x)) levels(x) else x)
 }
@@ -47,8 +47,8 @@ sample_fct <- function(x, N, ...) { # nolint
 #' random.cdisc.data:::relvar_init("Alanine Aminotransferase Measurement", "ALT")
 #' random.cdisc.data:::relvar_init("Alanine Aminotransferase Measurement", "U/L")
 relvar_init <- function(relvar1, relvar2) {
-  stopifnot(is_character_vector(relvar1))
-  stopifnot(is_character_vector(relvar2))
+  checkmate::assert_character(relvar1, min.len = 1, any.missing = FALSE)
+  checkmate::assert_character(relvar2, min.len = 1, any.missing = FALSE)
 
   if (length(relvar1) != length(relvar2)) {
     message(simpleError(
@@ -102,9 +102,9 @@ relvar_init <- function(relvar1, relvar2) {
 #'
 rel_var <- function(df = NULL, var_name = NULL, var_values = NULL, related_var = NULL) {
   stopifnot(is.null(df) || is.data.frame(df))
-  stopifnot(is.null(var_name) || is_character_single(var_name))
-  stopifnot(is.null(var_values) || is_character_vector(var_values))
-  stopifnot(is.null(related_var) || is_character_single(related_var))
+  checkmate::assert_string(var_name, null.ok = TRUE)
+  checkmate::assert_character(var_values, null.ok = TRUE, min.len = 1, any.missing = FALSE)
+  checkmate::assert_string(related_var, null.ok = TRUE)
 
   if (is.null(df)) {
     message("Missing data frame argument value.")
@@ -145,9 +145,10 @@ rel_var <- function(df = NULL, var_name = NULL, var_values = NULL, related_var =
 visit_schedule <- function(visit_format = "WEEK",
                            n_assessments = 10L,
                            n_days = 5L) {
-  stopifnot(is_character_single(visit_format))
-  stopifnot(is_integer_single(n_assessments))
-  stopifnot(is_integer_single(n_days))
+
+  checkmate::assert_string(visit_format)
+  checkmate::assert_integer(n_assessments, len = 1, any.missing = FALSE)
+  checkmate::assert_integer(n_days, len = 1, any.missing = FALSE)
 
   # trap invalid assessment format
   if (!(toupper(visit_format) %in% c("WEEK", "CYCLE"))) {
@@ -239,9 +240,9 @@ var_relabel <- function(x, ...) {
 #'
 apply_metadata <- function(df, filename, add_adsl = TRUE, adsl_filename = "metadata/ADSL.yml") { # nolint
   stopifnot(is.data.frame(df))
-  stopifnot(is_character_single(filename))
-  stopifnot(is_logical_single(add_adsl))
-  stopifnot(is_character_single(adsl_filename))
+  checkmate::assert_string(filename)
+  checkmate::assert_flag(add_adsl)
+  checkmate::assert_string(adsl_filename)
 
   apply_type <- function(df, var, type) {
     if (is.null(type)) {
@@ -328,7 +329,7 @@ apply_metadata <- function(df, filename, add_adsl = TRUE, adsl_filename = "metad
 #' @export
 replace_na <- function(v, percentage = 0.05, seed = NULL) {
 
-  stopifnot(is_numeric_single(percentage))
+  checkmate::assert_numeric(percentage, len = 1, any.missing = FALSE)
   stopifnot(percentage >= 0 && percentage <= 1)
 
   if (percentage == 0) {